# contracts/openapi.geo.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Uni\u002B \u2014 M\u00F3dulo Geo",
    "description": "API REST do Sistema Unificado Unifesspa (Uni\u002B). Contrato can\u00F4nico V1 \u2014 ProblemDetails RFC 9457, pagina\u00E7\u00E3o cursor opaco, vendor MIME versioning. Toda string user-facing em pt-BR (ADR-0022).",
    "contact": {
      "name": "CTIC Unifesspa",
      "url": "https://developers.uniplus.unifesspa.edu.br",
      "email": "ctic@unifesspa.edu.br"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.uniplus.unifesspa.edu.br",
      "description": "Produ\u00E7\u00E3o"
    },
    {
      "url": "https://api.hml.uniplus.unifesspa.edu.br",
      "description": "Homologa\u00E7\u00E3o"
    }
  ],
  "paths": {
...
    "/api/cep/{cep}": {
      "get": {
        "tags": [
          "Cep"
        ],
        "parameters": [
          {
            "name": "cep",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CepResolvidoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CepResolvidoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CepResolvidoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CepResolvidoDto": {
        "required": [
          "cep",
          "tipo",
          "logradouro",
          "complemento",
          "bairro",
          "distrito",
          "cidade",
          "codigoIbge",
          "uf",
          "latitude",
          "longitude",
          "nivelResolucao",
          "origem"
        ],
        "type": "object",
        "properties": {
          "cep": {
            "type": "string"
          },
          "tipo": {
            "type": [
              "null",
              "string"
            ]
          },
          "logradouro": {
            "type": [
              "null",
              "string"
            ]
          },
          "complemento": {
            "type": [
              "null",
              "string"
            ]
          },
          "bairro": {
            "type": [
              "null",
              "string"
            ]
          },
          "distrito": {
            "type": [
              "null",
              "string"
            ]
          },
          "cidade": {
            "type": "string"
          },
          "codigoIbge": {
            "type": "string"
          },
          "uf": {
            "type": "string"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "nivelResolucao": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          },
          "alternativos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogradouroAlternativoDto"
            }
          },
          "_links": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "LogradouroAlternativoDto": {
        "required": [
          "tipo",
          "logradouro",
          "bairro",
          "distrito",
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "tipo": {
            "type": [
              "null",
              "string"
            ]
          },
          "logradouro": {
            "type": "string"
          },
          "bairro": {
            "type": [
              "null",
              "string"
            ]
          },
          "distrito": {
            "type": [
              "null",
              "string"
            ]
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Token JWT emitido pelo issuer/realm OIDC configurado para a Geo API.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}